poc/pyproject.toml
[project]
name = "ai-supervision-poc"
version = "0.1.0"
description = "AI Supervision PoC - Audio transcription + Screen capture"
requires-python = ">=3.11"
dependencies = [
    "faster-whisper>=1.0.0",      # Whisper 转录 (CTranslate2 加速)
    "sounddevice>=0.4.6",          # 音频采集
    "numpy>=1.24.0",               # 数值计算
    "mss>=9.0.0",                  # 屏幕截图
    "pillow>=10.0.0",              # 图像处理
    "imagehash>=4.3.0",            # 图像哈希 (相似度检测)
]

[project.optional-dependencies]
vad = [
    "webrtcvad>=2.0.10",           # 语音活动检测
]

[project.scripts]
poc = "src.main:main"

[tool.uv]
dev-dependencies = [
    "ipython>=8.0.0",
]
